:test_tube: test (dot): Move test file.
